requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dill==0.3.1.1
Flask==1.1.1
idna==2.8
imgurpython==1.1.7
itsdangerous==1.1.0
Jinja2==2.10.3
klepto==0.1.8
MarkupSafe==1.1.1
Pillow==6.2.1
pox==0.2.7
pycodestyle==2.5.0
pykakasi==1.2
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0